chore: clean up unused components and improve error handling
- Removed obsolete action handling code.
- Updated coordinate types for consistency.
- Refactored error handling to use a unified error type.
- Simplified API response handling by removing unnecessary string conversions.
- Enhanced OCR functions to return more specific results.
```

+++ src-tauri/src/composent/window.rs
@@ -86,7 +86,7 @@ impl WindowManager {
         }
     }
 
-    pub fn send_travel_command(&self, x: i32, y: i32) -> Result<()> {
+    pub fn send_travel_command(&self, x: i8, y: i8) -> Result<()> {
         let settings = Settings::default();
         let mut enigo = Enigo::new(&settings)?;
         let command = format!("/travel {},{}", x, y);
